Add tests for populate_board! and build that function
The tests just check for the correct number of chess pieces regardless
of color and basic default attributes.
The populate_board! function is currently storing arrays which will
be changed soon by removing the piece table's square_location field
and replacing it with a x_position and y_position fields. Then I'll
update the function as needed to work with that.