Replace square_location with x and y position in the piece table
------------------------------------------------------------------------
The 2 migration files remove the square_location column in the pieces
table and then add x_position and y_position columns. This was done
because square_location was previously storing an array to hold a chess
piece's location which was not ideal because it turns out that querying
a database for an array is difficult. Yay, for simplicity.