Build tests for testing default positions of pieces
------------------------------------------------------------------------
I added additional tests for checking that initial pieces locations
match the correct initial positions. May have went a little overboard
on testing. Note: This commit has a failing test. This particular test
is supposed to check that a game's generated chess pieces are
associated with the game's id. For whatever reason, the test fails; and
I have a suspicion that I've overlooked something basic. Let me know if
you find what it is.