Make the obstacle method in piece.rb a normal method
------------------------------------------------------------------------
Earlier the obstacle method was inside the is_obstructed method. This
means that it was a method inside another method which is not exactly
clear way to build a function, and it wasn't able to be used by
potential future functions. So now I moved it outside of is_obstructed
as a normal method.