Make move_to! call valid_move? and add extra test for it
------------------------------------------------------------------------
The thinking behind this is that after the app receives a new
destination from a chess player, the app will run move_to! which will
first guard against an invalid move or obstacle. Assuming the move is
valid and no obstacle exists, move_to! will procede with its remaining
logic.