Finalized validate_move to call valid_move? and Bug fixes for view
Changed method name to validate_move which will call valid_move? for each piece
Corrected the bugs in view: Changed square link to only visible after piece link
has been selected. This makes the square not clickable if a piece is not selected
thus fixing the bug that caused a redirect to an alternate game when an empty
square was clicked prior to selecting a piece to move.